requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
backoff==2.2.1
boto3==1.34.34
botocore==1.34.162
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
fastapi-sso==0.16.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.72.6.post1
litellm-enterprise==0.1.7
litellm-proxy-extras==0.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.3
mdurl==0.1.2
multidict==6.5.0
nodeenv==1.9.1
oauthlib==3.3.0
openai==1.88.0
orjson==3.10.18
packaging==25.0
prisma==0.15.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.18
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==13.7.1
rpds-py==0.25.1
rq==2.4.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
websockets==13.1
yarl==1.20.1
zipp==3.23.0
